backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pydantic==2.4.2
schedule==1.2.0
slack-sdk==3.21.3
sendgrid==6.10.0
twilio==8.8.0
pandas==2.1.1
numpy==1.24.3
sqlalchemy==2.0.21
alembic==1.12.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0 